+++ tests/manual/6-pageSpinner.R
@@ -1,0 +1,17 @@
+library(shiny)
+
+ui <- fluidPage(
+  actionButton("go", "Go"),
+  plotOutput("plot")
+)
+server <- function(input, output) {
+  observeEvent(input$go, {
+    showPageSpinner()
+    Sys.sleep(1)
+    hidePageSpinner()
+  })
+  output$plot <- renderPlot({
+    plot(runif(10))
+  })
+}
+shinyApp(ui, server)